Modifies sample uploading form
