+++ src/views/samples/CreateSampleForm.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { SampleRepository } from '@/repositories/SampleRepository';
-import router from '@/router';
 import Button from '@/shadcn/ui/button/Button.vue';
 import Dialog from '@/shadcn/ui/dialog/Dialog.vue';
 import DialogClose from '@/shadcn/ui/dialog/DialogClose.vue';
@@ -16,24 +15,35 @@ import { useAuthStore } from '@/stores/auth';
 import { useUserStore } from '@/stores/user';
 import { ref, type Ref } from 'vue';
 import UploadFile from './UploadFile.vue';
-import ConfigureAlignment from './ConfigureAlignment.vue';
+import Switch from '@/shadcn/ui/switch/Switch.vue';
+import Select from '@/shadcn/ui/select/Select.vue';
+import SelectTrigger from '@/shadcn/ui/select/SelectTrigger.vue';
+import SelectValue from '@/shadcn/ui/select/SelectValue.vue';
+import SelectContent from '@/shadcn/ui/select/SelectContent.vue';
+import SelectGroup from '@/shadcn/ui/select/SelectGroup.vue';
+import SelectItem from '@/shadcn/ui/select/SelectItem.vue';
+import router from '@/router';
+import type { Group } from './types';
 import type { ProjectFile } from '@/repositories/types/entities/ProjectFile';
-
+import { AnalyzerRepository } from '@/repositories/AnalyzerRepository';
 
 const authStore = useAuthStore();
 const userStore = useUserStore();
 
 // Repositories
 const sampleRepository: SampleRepository = new SampleRepository();
+const analyzerRepository: AnalyzerRepository = new AnalyzerRepository();
 const new_sample_name: Ref<string> = ref('');
 const new_sample_comment: Ref<string> = ref('');
 const new_sample_tags: Ref<string> = ref('');
+const chemistry: Ref<string> = ref('');
+const genome: Ref<string> = ref('');
 
 // Refs
-const files_uploaded: Ref<Array<ProjectFile>> = ref([])
 const step: Ref<number> = ref(0)
 const sample_id: Ref<string> = ref('')
-const file_type: Ref<string> = ref('')
+const fastq: Ref<boolean> = ref(true)
+const files_uploaded: Ref<Array<ProjectFile>> = ref([])
 
 async function newSample() {
     const res = await sampleRepository.createSample({
@@ -41,7 +51,7 @@ async function newSample() {
         data: {
             name: new_sample_name.value,
             description: new_sample_comment.value,
-            tags: new_sample_tags.value.replace(" ","").split(',')
+            tags: new_sample_tags.value.replace(" ", "").split(',')
         },
         orgId: userStore.defaultOrg?.id ?? ''
     });
@@ -49,9 +59,72 @@ async function newSample() {
     step.value++;
 }
 
-const submitForm = () => {
-    document.getElementById('submitButton')?.click();
+async function align() {
+    let platform: string = "";
+    
+    if (fastq) {
+        platform = chemistry.value.includes("10x") ? "10x" : "hydrop";
+    }
+
+    let analyzer_name = 'fastq_initialization';
+    if (!fastq) {
+        analyzer_name = 'h5_initialization';
+    }
+
+    const analyzer = await analyzerRepository.getAnalyzerByName({
+        bearerToken: authStore.getToken ?? '',
+        orgId: userStore.defaultOrg?.id ?? '',
+        analyzer_name: analyzer_name
+    });
+
+    let files = [];
+    for (const file of files_uploaded.value){
+        files.push(file.name)
+    }
+
+    const group:Group = {
+        name: "group",
+        files: files
+    }
+    await sampleRepository.createAnalysis({
+        orgId: userStore.defaultOrg?.id ?? '',
+        sampleId: sample_id.value,
+        bearerToken: authStore.getToken ?? '',
+        handleBusinessErrors: true,
+        data: {
+            analyzer_id: analyzer.data.id,
+            config: {
+                r: {
+                    sample: sample_id.value,
+                    type: analyzer_name
+                },
+                fastqc: {
+                    sample: sample_id.value,
+                },
+                star: {
+                    sample: sample_id.value,
+                    genome: genome.value,
+                    whitelist: chemistry.value,
+                    platform: platform,
+                    groups: [group]
+                },
+                scanpy: {
+                    sample: sample_id.value,
+                    file_type: fastq? 'fastq' : 'h5'
+                },
+                fastp: {
+                    platform: platform,
+                    sample: sample_id.value,
+                },
+            },
+            branch: ' ', // This will be removed
+            commit_hash: ' ' // This will be removed
+        }
+    });
+
+    router.go(0)
 };
+
 </script>
 
 <template>
@@ -61,6 +134,8 @@ const submitForm = () => {
                 + New sample
             </Button>
         </DialogTrigger>
+
+        <!-- SAMPLE CREATION FORM -->
         <DialogContent v-if="step == 0" class="sm:max-w-[425px]">
             <DialogHeader>
                 <DialogTitle>Add a new sample</DialogTitle>
@@ -71,18 +146,75 @@ const submitForm = () => {
             <div class="grid gap-4 py-4">
                 <div class="grid grid-cols-4 items-center gap-4">
                     <Label for="name" class="text-right"> Name </Label>
-                    <Input id="name" placeholder="Sample XYZ" class="col-span-3" v-model="new_sample_name"
-                        @keyup.enter="submitForm" />
+                    <Input id="name" placeholder="Sample XYZ" class="col-span-3" v-model="new_sample_name" />
                 </div>
                 <div class="grid grid-cols-4 items-center gap-4">
                     <Label for="description" class="text-right"> Comment </Label>
-                    <Input id="description" placeholder="Your comment." class="col-span-3" v-model="new_sample_comment"
-                        @keyup.enter="submitForm" />
+                    <Input id="description" placeholder="Your comment." class="col-span-3"
+                        v-model="new_sample_comment" />
                 </div>
                 <div class="grid grid-cols-4 items-center gap-4">
                     <Label for="tags" class="text-right"> Tags </Label>
-                    <Input id="tags" placeholder="Comma, separated, tags" class="col-span-3" v-model="new_sample_tags"
-                        @keyup.enter="submitForm" />
+                    <Input id="tags" placeholder="Comma, separated, tags" class="col-span-3"
+                        v-model="new_sample_tags" />
+                </div>
+                <div class="grid grid-cols-4 items-center gap-4">
+                    <Label for="file-type" class="text-right">File Type</Label>
+                    <div class="flex gap-2 items-center justify-center text-xs col-span-3">
+                        <span>H5</span>
+                        <Switch id="file-type" class="col-span-2" :checked="fastq" @update:checked="fastq = !fastq" />
+                        <span>FASTQ</span>
+                    </div>
+                </div>
+                <div v-if="fastq">
+                    <Label for="file-type" class="text-right">Select Chemistry</Label>
+                    <Select v-model="chemistry" class="col-span-3">
+                        <SelectTrigger>
+                            <SelectValue placeholder="Select the relevant chemistry" />
+                        </SelectTrigger>
+                        <SelectContent>
+                            <!-- https://kb.10xgenomics.com/hc/en-us/articles/115004506263-What-is-a-barcode-whitelist -->
+                            <SelectGroup>
+                                <SelectItem value="10x-2014">
+                                    Single Cell 3' v1 (2014)
+                                </SelectItem>
+                                <SelectItem value="10x-2016">
+                                    Single Cell 3' v2, Single Cell 5' v1 and v2, Single Cell 5' HT v2 (2016)
+                                </SelectItem>
+                                <SelectItem value="10x-2018">
+                                    Single Cell 3' v3, Single Cell 3' v3.1, Single Cell 3' HT v3.1 (2018)
+                                </SelectItem>
+                            </SelectGroup>
+                        </SelectContent>
+                    </Select>
+                </div>
+                <div v-if="fastq">
+                    <Label for="file-type" class="text-right">Select Genome</Label>
+                    <Select v-model="genome">
+                        <SelectTrigger>
+                            <SelectValue placeholder="Select the relevant genome" />
+                        </SelectTrigger>
+                        <SelectContent>
+                            <SelectGroup>
+                                <!-- <SelectItem value="gencodegenes_GRCm39_mouse">
+                                    Mouse GRCm39 GencodeGenes
+                                </SelectItem>
+                                <SelectItem value="nih_GRCm39_mouse">
+                                    Mouse GRCm39 NIH
+                                </SelectItem>
+                                <Separator></Separator> -->
+                                <SelectItem value="10x_GRCh38_human">
+                                    Human GRCh38 10x (Recommended)
+                                </SelectItem>
+                                <SelectItem value="nih_GRCh38_human">
+                                    Human GRCh38 NIH
+                                </SelectItem>
+                                <SelectItem value="gencodegenes_GRCh38_human">
+                                    Human GRCh38 GencodeGenes
+                                </SelectItem>
+                            </SelectGroup>
+                        </SelectContent>
+                    </Select>
                 </div>
             </div>
             <DialogFooter>
@@ -91,6 +223,8 @@ const submitForm = () => {
                 </Button>
             </DialogFooter>
         </DialogContent>
+
+        <!-- FILE UPLOAD FORM -->
         <DialogContent v-else-if="step == 1" class="sm:max-w-[425px]">
             <DialogHeader>
                 <DialogTitle>Upload your files</DialogTitle>
@@ -98,38 +232,17 @@ const submitForm = () => {
                     Import FASTQs or H5 files for this sample.
                 </DialogDescription>
             </DialogHeader>
-            <UploadFile v-model:file_type="file_type" v-model:sample_id="sample_id" v-model:files_uploaded="files_uploaded"></UploadFile>
-            <DialogFooter>
-
-                <Button class="rounded-full" @click="step--">
-                    Previous
-                </Button>
-                <Button v-if="file_type = 'fastq'" class="rounded-full" @click="step++">
-                    Next
-                </Button>
-                <DialogClose v-else as-child>
-                    <Button class="rounded-full">
-                        Finish
-                    </Button>
-                </DialogClose>
-
-            </DialogFooter>
-        </DialogContent>
-        <DialogContent v-else-if="step == 2 && file_type == 'fastq'" class="sm:max-w-[425px]">
-            <DialogHeader>
-                <DialogTitle>Configure the alignment</DialogTitle>
-                <DialogDescription>
-                    Select a Genome and the chemistry.
-                </DialogDescription>
-            </DialogHeader>
-            <ConfigureAlignment v-model:step="step" :sample_id="sample_id" :file_type="file_type" :files_uploaded="files_uploaded"></ConfigureAlignment>
+            <UploadFile :sample_id="sample_id" v-model:files_uploaded="files_uploaded" :align="align"/>
             <DialogFooter>
                 <Button class="rounded-full" @click="step--">
                     Previous
                 </Button>
                 <DialogClose as-child>
-                    <Button class="rounded-full" type="submit" form="alignmentForm">
+                    <Button v-if="fastq" class="rounded-full" @click="align">
                         Align
+                    </Button>
+                    <Button v-else class="rounded-full">
+                        Close
                     </Button>
                 </DialogClose>
             </DialogFooter>
